+++ index.js
@@ -109,22 +109,40 @@ class Jetpack {
     return this.__options;
   }
 
-  // Function overrides, etc.
-  _functionOptions({ functionObject }) {
-    if (!functionObject) {
+  // Function, layer overrides, etc.
+  // eslint-disable-next-line complexity
+  _extraOptions({ functionObject, layerObject }) {
+    if (!functionObject && !layerObject) {
       return this._serviceOptions;
     }
 
     const opts = Object.assign({}, this._serviceOptions);
-    const fnOpts = functionObject.jetpack || {};
-    if (fnOpts.roots) {
-      opts.roots = (opts.roots || []).concat(fnOpts.roots);
+    const fnRoots = ((functionObject || {}).jetpack || {}).roots;
+    const layerRoots = ((layerObject || {}).jetpack || {}).roots;
+    if (fnRoots || layerRoots) {
+      opts.roots = (opts.roots || [])
+        .concat(fnRoots || [])
+        .concat(layerRoots || []);
     }
 
     return opts;
   }
 
-  filePatterns({ functionObject }) {
+  // Helper for layer patterns we'll need for all packages
+  get _layerExcludes() {
+    if (this.__layerExcludes) { return this.__layerExcludes; }
+
+    const { service } = this.serverless;
+    this.__layerExcludes = service.getAllLayers()
+      .map((layer) => service.getLayer(layer))
+      .filter((layerObj) => layerObj.path)
+      .map((layerObj) => `${layerObj.path}/**`);
+
+    return this.__layerExcludes;
+  }
+
+  // eslint-disable-next-line complexity,max-statements
+  filePatterns({ functionObject, layerObject }) {
     const { service, pluginManager } = this.serverless;
     const servicePackage = service.package;
     const serviceInclude = servicePackage.include || [];
@@ -134,9 +152,19 @@ class Jetpack {
     const functionInclude = functionPackage.include || [];
     const functionExclude = functionPackage.exclude || [];
 
+    const layerPackage = (layerObject || {}).package || {};
+    const layerInclude = layerPackage.include || [];
+    const layerExclude = layerPackage.exclude || [];
+
     // Combined, unique patterns, in stable sorted order (remove _later_ instances).
     // This is `_.union` in serverless built-in.
-    const include = union(serviceInclude, functionInclude);
+    let include = serviceInclude;
+    if (functionInclude) {
+      include = union(include, functionInclude);
+    }
+    if (layerInclude) {
+      include = union(include, layerInclude);
+    }
 
     // Packaging logic.
     //
@@ -144,8 +172,10 @@ class Jetpack {
     // - Default excludes
     // - Exclude serverless config files
     // - Exclude plugin local paths
+    // - Exclude all layers
     // - Apply service package excludes
     // - Apply function package excludes
+    // - Apply layer package excludes
     //
     // https://serverless.com/framework/docs/providers/aws/guide/packaging#exclude--include
     // > At first it will apply the globs defined in exclude. After that it'll
@@ -173,7 +203,9 @@ class Jetpack {
       slsDefaultExcludePatterns,
       pluginsLocalPath ? [pluginsLocalPath] : null,
       serviceExclude,
-      functionExclude
+      this._layerExcludes,
+      functionExclude,
+      layerExclude
     ]
       .filter((arr) => !!arr && arr.length)
       .reduce((memo, arr) => union(memo, arr), []);
@@ -184,19 +216,22 @@ class Jetpack {
     };
   }
 
-  async globAndZip({ bundleName, functionObject, worker }) {
+  async globAndZip({ bundleName, functionObject, layerObject, worker }) {
     const { config } = this.serverless;
     const servicePath = config.servicePath || ".";
-    const { base, roots } = this._functionOptions({ functionObject });
-    const { include, exclude } = this.filePatterns({ functionObject });
+    const layerPath = (layerObject || {}).path;
+    const cwd = layerPath ? path.relative(servicePath, layerPath) : servicePath;
+
+    const { base, roots } = this._extraOptions({ functionObject, layerObject });
+    const { include, exclude } = this.filePatterns({ functionObject, layerObject });
 
     const buildFn = worker ? worker.globAndZip : globAndZip;
     const { numFiles, bundlePath, buildTime } = await buildFn(
-      { servicePath, base, roots, bundleName, include, exclude }
+      { cwd, servicePath, base, roots, bundleName, include, exclude }
     );
 
     this._logDebug(
-      `Zipped ${numFiles} sources from ${servicePath} to artifact location: ${bundlePath}`
+      `Zipped ${numFiles} sources from ${cwd} to artifact location: ${bundlePath}`
     );
 
     return { numFiles, bundlePath, buildTime };
@@ -237,7 +272,21 @@ class Jetpack {
     this._log(`Packaged service: ${bundleName} (${toSecs(buildTime)}s)`);
   }
 
-  // eslint-disable-next-line max-statements
+  async packageLayer({ layerName, layerObject, worker }) {
+    const bundleName = path.join(SLS_TMP_DIR, `${layerName}.zip`);
+
+    // Package.
+    this._logDebug(`Start packaging layer: ${bundleName}`);
+    const { buildTime } = await this.globAndZip({ bundleName, layerObject, worker });
+
+    // Mutate serverless configuration to use our artifacts.
+    layerObject.package = layerObject.package || {};
+    layerObject.package.artifact = bundleName;
+
+    this._log(`Packaged layer: ${bundleName} (${toSecs(buildTime)}s)`);
+  }
+
+  // eslint-disable-next-line max-statements,complexity
   async package() {
     const { service } = this.serverless;
     const servicePackage = service.package;
@@ -246,13 +295,16 @@ class Jetpack {
     let tasks = [];
     let worker;
 
+    // ------------------------------------------------------------------------
+    // Lambdas
+    // ------------------------------------------------------------------------
     // Check if we have a single function limitation from `deploy -f name`.
     const singleFunctionName = (this.options || {}).function;
     if (singleFunctionName) {
       this._logDebug(`Packaging only for function: ${singleFunctionName}`);
     }
 
-    // Gather internal configuration.
+    // Functions.
     const fnsPkgs = service.getAllFunctions()
       // Limit to single function if provided.
       .filter((functionName) => !singleFunctionName || singleFunctionName === functionName)
@@ -298,10 +350,40 @@ class Jetpack {
       this._logDebug("No matching service or functions to package.");
     }
 
+    // ------------------------------------------------------------------------
+    // Layers
+    // ------------------------------------------------------------------------
+    const layersPkgs = service.getAllLayers()
+      // Convert to more useful format.
+      .map((layerName) => ({
+        layerName,
+        layerObject: service.getLayer(layerName)
+      }))
+      .map((obj) => ({
+        ...obj,
+        layerPackage: obj.layerObject.package || {}
+      }))
+      .map((obj) => ({
+        ...obj,
+        disable: !!obj.layerObject.disable,
+        artifact: obj.layerObject.artifact
+      }));
+
+    // Package layers if not in `-f NAME` single-function mode.
+    let numLayers = 0;
+    if (!singleFunctionName) {
+      // Get list of layers to package.
+      const layersPkgsToPackage = layersPkgs.filter((obj) => !(obj.disable || obj.artifact));
+      numLayers = layersPkgsToPackage.length;
+      tasks = tasks.concat(layersPkgsToPackage.map((obj) => () =>
+        this.packageLayer({ ...obj, worker })
+      ));
+    }
+
     // Run all packaging work.
     this._log(
-      `Packaging ${numFns} functions and ${shouldPackageService ? 1 : 0} services `
-      + `with concurrency ${concurrency}`
+      `Packaging ${numFns} functions, ${shouldPackageService ? 1 : 0} services, and `
+      + `${numLayers} layers with concurrency ${concurrency}`
     );
     if (concurrency > 1) {
       // Run concurrently.