+++ test/packages/complex/npm/layers/no-deps/index.js
@@ -1,0 +1,5 @@
+"use strict";
+
+exports.module = {
+  msg: "I have no dependencies"
+};